[Refactor] utils 폴더 생성 -> 함수 관리
