+++ src/pages/flame/lineup/FlameLineupPage.jsx
@@ -5,30 +5,13 @@ import 'aos/dist/aos.css';
 import InstalogoImage from '@/assets/webps/wdfLineup/instagramLogo.webp';
 import testImage from '@/assets/webps/wdfLineup/test.webp';
 import * as S from './styles'; // 스타일 파일에서 스타일 컴포넌트를 가져옴
+import { getSelectedDayByDate } from '@/utils/stage/getSelectedDayByDate';
 
 const FlameLineupPage = () => {
   const [selectedDay, setSelectedDay] = useState('Day1');
 
   useEffect(() => {
-    // 현재 날짜 가져오기
-    const today = new Date();
-    const month = today.getMonth() + 1; // 월 (0부터 시작하므로 1 더함)
-    const day = today.getDate(); // 일
-
-    // 날짜에 따른 Day 선택
-    if (month === 9) {
-      if (day === 25) {
-        setSelectedDay('Day1');
-      } else if (day === 26) {
-        setSelectedDay('Day2');
-      } else if (day === 27) {
-        setSelectedDay('Day3');
-      } else {
-        setSelectedDay('Day1'); // 9월 25-27일이 아닌 경우 Day1으로 설정
-      }
-    } else {
-      setSelectedDay('Day1'); // 9월이 아닌 경우 Day1으로 설정
-    }
+    setSelectedDay(getSelectedDayByDate());
 
     AOS.init({
       duration: 1000, // 애니메이션 지속 시간 (밀리초)
